Improves TransformType Consistency
In some places TransformType was being bound to T as a generic type, in others it was being bound to Transform.
This causes some readability concerns owing to the usage of T as a generic type for other contexts within the same file.
This commit replaces instances of the former with the latter.